// app/models/Settings.ts
class Settings {
    name: string;
    isCurrent: boolean;
    flashcards: any;
    whiteboard: any;
    timedReading: any;
    wordcards: any;
    lettercards: any;
    blackboard: any;
    blankTiles: any;
    workbookTiles: any;
    miniTiles: any;
    spelling: any;
    sentences: any;
    phrases: any;
    passages: any;
    pdfViewer: any;
    ownerKey: string;
    tilefont: string;
    textfont: string;
    bgcolor: string;
    tilehighlightcolor: string;
    whizzimap: boolean;
    wordlists: any;
    students: any;
    compoundSounds: any;
    tiles: any;
    _id: string;
    bank: any;
    fontSize: number;
    themeHighlightColor: string;
    snap: boolean;
    proper: boolean;
    tileUpdating: number;
    wordUpdating: number;
}

export {Settings}